+++ src/main/scala/com/peoplepattern/text/package.scala
@@ -1,0 +1,13 @@
+package com.peoplepattern
+
+/**
+ * Text processing utilities for language prediction, tokenization, term extraction etc
+ *
+ * <ul>
+ * <li>See [[LangBundle]] for tokenization and term extraction
+ * <li>See [[LanguageIdentifier]] for language identification
+ * <li>See [[StringUtil]] for low-level string helpers
+ * <li>See [[Implicits.StringWithAnalysis]] for implicit helper methods for strings
+ * </ul>
+ */
+package object text